move wxmap gui to main larch lib
