+++ larch/wxmap/mapimageframe.py
