--- plugins/wx/__init__.py
@@ -1,5 +1,0 @@
-from larch.wxlib import (gui_utils, ParameterPanel, PeriodicTablePanel,
-                         _newplot, _plot, XRFDisplayFrame)
-
-from .gse_dtcorrect import DTViewer
-from .mapviewer import MapViewer
