--- plugins/wx/requirements.txt
@@ -1,1 +1,0 @@
-wx>=2.8.12
